services/orchestration_engine/requirements.txt
Fflask
flask-cors
celery
redis
requests
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
httpx
python-multipart
aiokafka
Flask-Cors
PyJWT
Werkzeug
